refactored code for consistent naming standards, reorganized generate snapshot code, fixed pattern 2 snapshot logic
