+++ cdc-from-snapshot-demo/README.md
@@ -1,5 +1,8 @@
+# DLT APPLY CHANGES FROM SNAPSHOT
 ```bash
 cd cdc-from-snapshot
+databricks bundle validate
 databricks bundle deploy
-databricks bundle run single_snapshot_job
+databricks bundle run dlt_snapshot_ingestion_pattern1_job
+databricks bundle run dlt_snapshot_ingestion_pattern2_job
 ```
