+++ cdc-from-snapshot-demo/src/notebooks/order_count_by_status_pattern1.sql
