+++ cdc-from-snapshot-demo/src/notebooks/process_snapshot_pattern1.py
